Trim itfMesManager 003a5ca0 parameters

+++ src/Event/mt_evtMisc.c
@@ -731,7 +731,7 @@ void FUN_0038b960(int param_1,long param_2,short param_3);
 void FUN_0038bc70(int param_1,int param_2,int param_3);
 void FUN_0038c460(long param_1,int param_2,int *param_3,u32 *param_4);
 void FUN_0038c540(int param_1,long param_2,int param_3,u32 *param_4, u32 *param_5,u32 *param_6);
-void FUN_0038c830(int param_1,long param_2);
+void FUN_0038c830(int param_1,int param_2);
 void FUN_0038ca00(void);
 void FUN_0038ca80(int param_1);
 void FUN_0038cc10(int param_1,u16 *param_2,long param_3);
@@ -5885,7 +5885,7 @@ void FUN_0038c540(int param_1,long param_2,int param_3,u32 *param_4,
 // FUN_0038C830 NONMATCHING
 
 
-void FUN_0038c830(int param_1,long param_2)
+void FUN_0038c830(int param_1,int param_2)
 {
   u16 uVar1;
   bool bVar2;
